Lab 2 - GitHub Labs/requirements.txt
pandas
scikit-learn
google-cloud-storage
pytest
joblib
python-dotenv
